Move Widget class to it's own file
